# src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeSkills Agent API",
    "version": "1.0.0",
    "description": "API for AI agents to create secrets, authenticate, and use skills (EVM wallet, Polymarket, raw signer, data sources).\n\n## Getting Started\n\n1. **Create a secret** via `POST /api/secrets` (no auth required). This returns an API key (`ssk_xxx`) and a claim URL.\n2. **Use the API key** as a Bearer token in the `Authorization` header for all skill endpoints.\n3. **Give the claim URL** to the secret owner so they can claim ownership, set policies, and manage the secret from the dashboard.\n\n## Re-linking\n\nIf an agent loses its API key, the secret owner can generate a one-time re-link token from the dashboard. The agent then calls `POST /api/secrets/relink` with that token to obtain a new API key.\n\n## Skill Execution Results\n\nSkill endpoints that perform on-chain actions return a `SkillExecutionResult` with one of three statuses:\n- **`executed`** (200) — Action completed successfully.\n- **`denied`** (403) — Blocked by a policy rule. The `reason` field explains why.\n- **`pending_approval`** (202) — Requires manual approval from the secret owner."
  },
  "servers": [
    {
      "url": "https://safeskills.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key issued when a secret is created (`ssk_xxx`). Pass as `Authorization: Bearer ssk_xxx`."
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {}
            }
          }
        }
      },
      "SecretType": {
        "type": "string",
        "enum": ["EVM_WALLET", "POLYMARKET_WALLET", "RAW_SIGNER", "API_KEY", "SSH_KEY", "OAUTH_TOKEN", "DATA_SOURCES"],
        "description": "Type of secret to create. Determines which skill endpoints are available."
      },
...
      "Secret": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "$ref": "#/components/schemas/SecretType" },
          "memo": { "type": "string" },
          "userId": { "type": "string", "format": "uuid", "nullable": true, "description": "Set after the owner claims the secret" },
          "smartAccountAddress": { "type": "string", "nullable": true, "description": "For EVM_WALLET / POLYMARKET_WALLET types" },
          "eoaAddress": { "type": "string", "nullable": true, "description": "For EVM_WALLET / POLYMARKET_WALLET types" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/api/secrets": {
      "post": {
        "operationId": "createSecret",
        "tags": ["Authentication"],
        "summary": "Create a new secret and get an API key",
        "description": "No authentication required. Creates a secret of the specified type and returns an API key for agent access.\n\nFor `EVM_WALLET` and `POLYMARKET_WALLET` types, a private key and smart account are generated automatically. For `RAW_SIGNER`, a signing seed is generated. For other types (`API_KEY`, `SSH_KEY`, `OAUTH_TOKEN`), the owner sets the value later via the dashboard.\n\nThe returned `claimUrl` should be given to the secret owner so they can claim ownership.\n\nRate limited to 5 requests per IP per 15 minutes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type"],
                "properties": {
                  "type": { "$ref": "#/components/schemas/SecretType" },
                  "memo": { "type": "string", "maxLength": 500, "description": "Optional memo describing the secret's purpose" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Secret created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "secret": { "$ref": "#/components/schemas/Secret" },
                            "apiKey": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string", "format": "uuid" },
                                "key": { "type": "string", "description": "Plain API key (`ssk_xxx`). Only returned once — store it securely." }
                              }
                            },
                            "claimUrl": { "type": "string", "format": "uri", "description": "URL for the owner to claim this secret" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": { "description": "Rate limited (5 per IP per 15 min)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/secrets/info": {
      "get": {
        "operationId": "getSecretInfo",
        "tags": ["Authentication"],
        "summary": "Get secret info for the current API key",
        "description": "Returns the secret associated with the provided API key. Useful for checking secret type, wallet addresses, and claim status.",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Secret info",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    { "type": "object", "properties": { "data": { "type": "object", "properties": { "secret": { "$ref": "#/components/schemas/Secret" } } } } }
                  ]
                }
              }
            }
          },
          "401": { "description": "Invalid or missing API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/secrets/relink": {
      "post": {
        "operationId": "relinkSecret",
        "tags": ["Authentication"],
        "summary": "Obtain a new API key via re-link token",
        "description": "No authentication required — the re-link token acts as proof of authorization.\n\nThe secret owner generates a one-time re-link token from the dashboard (valid for 10 minutes). The agent uses that token here to get a fresh API key.\n\nThe token is consumed on use and cannot be reused.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["relinkToken"],
                "properties": {
                  "relinkToken": { "type": "string", "minLength": 1, "description": "One-time re-link token from the secret owner" },
                  "apiKeyName": { "type": "string", "maxLength": 100, "description": "Optional name for the new API key" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New API key issued",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "secret": { "$ref": "#/components/schemas/Secret" },
                            "apiKey": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string", "format": "uuid" },
                                "key": { "type": "string", "description": "Plain API key (`ssk_xxx`). Only returned once." }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": { "description": "Invalid or expired re-link token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}